fixup: Add WorldtubeConstraint
