+++ src/Evolution/Systems/Cce/CMakeLists.txt
@@ -21,6 +21,7 @@ spectre_target_sources(
   ScriPlusValues.cpp
   SpecBoundaryData.cpp
   WorldtubeBufferUpdater.cpp
+  WorldtubeConstraint.cpp
   WorldtubeDataManager.cpp
   WorldtubeModeRecorder.cpp
   )
@@ -53,6 +54,7 @@ spectre_target_headers(
   KleinGordonSystem.hpp
   Tags.hpp
   WorldtubeBufferUpdater.hpp
+  WorldtubeConstraint.hpp
   WorldtubeDataManager.hpp
   WorldtubeModeRecorder.hpp
   )
